==> R/2.3-analysis.rmd <==
This file generates all plots and tables for final analysis and visualization 
of the projects results.

```{r, libraries}
library(ggplot2)
library(ggpubr)
library(viridis)
library(dplyr)
library(terra)
library(tidyterra)
library(rnaturalearth)
library(sf)
library(factoextra)
library(magick)
library(png)
library(grid)
library(gridExtra)
source("R/0.0-functions.r", encoding = "UTF-8") # self written functions used
```

```{r}
tot_time <- Sys.time()
```

```{r}
# used colour hex codes:
c("#00ba38", "#f8766d", "#3f80f1")
ftheme <- theme(
    legend.title = element_text(size = 15), legend.key.size = unit(1, "cm"),
    legend.text = element_text(size = 14), axis.title = element_text(size = 14)
)
```

```{r, plot presence yearly count 2002-2022}
occs <- readRDS("R/data/occurrence_data/axyridis_clean.rds")
cat("presences before 2002:", nrow(subset(occs, Year < 2002)), "\n")
occs <- subset(occs, Year >= 2002)
data_eu <- as.data.frame(table(subset(occs, Area == "eu")$Year))
data_eu$Area <- "eu"
data_as <- as.data.frame(table(subset(occs, Area == "as")$Year))
data_as$Area <- "as"
data <- merge(data_eu, data_as, by = "Var1", all.x = TRUE)
data[is.na(data)] <- 0
data$Area.y <- "as"
data_eu <- data[, c(1, 2, 3)]
data_as <- data[, c(1, 4, 5)]
names(data_eu) <- c("Year", "Freq", "Area")
names(data_as) <- names(data_eu)
data <- rbind(data_eu, data_as)

p <- ggplot(data, aes(x = Year, y = Freq, fill = Area)) +
    geom_point(aes(colour = Area, shape = Area), size = 2.5) +
    geom_line(aes(colour = Area, group = Area)) +
    scale_y_log10() +
    scale_color_manual(values = c("#00ba38", "#f8766d")) +
    scale_x_discrete(breaks = seq(2002, 2022, by = 4)) +
    labs(x = "Year", y = "Number of Observations") +
    theme_pubr() +
    ftheme
ggsave(p, width = 8, height = 5, filename = "R/figures/pres-per-year-log.png")
```

```{r, plot comparison raw dataset to cleaned}
# use equal earth crs? greyscale?
raw <- read.csv("R/data/occurrence_data/Harmonia-axyridis_gbif_raw.csv",
    header = TRUE,
    sep = "\t"
)
clean <- readRDS("R/data/occurrence_data/axyridis_clean.rds")
cat("raw:", nrow(raw), "|clean:", nrow(clean), "\n")
raw <- raw[!(raw$decimalLongitude %in% clean$Lon) & !(raw$decimalLatitude %in% clean$Lat), ]
lc_ref <- rast("R/data/cropped_rasters/Cop_LC_2002_eu.tif")

raw_v <- vect(raw, geom = c("decimalLongitude", "decimalLatitude"), crs = crs(lc_ref))
clean_v <- vect(clean, geom = c("Lon", "Lat"), crs = crs(lc_ref))
exts_v <- vect(ext(lc_ref), crs = crs(lc_ref))
lc_ref <- rast("R/data/cropped_rasters/Cop_LC_2002_as.tif")
exts_v <- rbind(exts_v, vect(ext(lc_ref), crs = crs(lc_ref)))
# countries map
countries <- ne_countries(scale = "medium", returnclass = "sf")
countries <- st_transform(countries, crs = crs(lc_ref))
countries <- vect(countries)

# plot
p <- ggplot(countries) +
    geom_spatvector(colour = "black", fill = "white") +
    geom_spatvector(data = raw_v, colour = "#3f80f1") +
    geom_spatvector(data = subset(clean_v, clean_v$Area == "eu"), colour = "#f8766d") +
    geom_spatvector(data = exts_v[1], colour = "#f8766d", fill = NA, linetype = "solid", linewidth = 1) +
    geom_spatvector(data = subset(clean_v, clean_v$Area == "as"), colour = "#00ba38") +
    geom_spatvector(data = exts_v[2], colour = "#00ba38", fill = NA, linetype = "solid", linewidth = 1) +
    scale_x_continuous(expand = c(0, 0)) +
    scale_y_continuous(expand = c(0, 0)) +
    theme_pubr()
ggsave(p, width = 8, height = 5, filename = "R/figures/raw-vs-cleaned-glob.png")
```

```{r, plot extent subdiv method and pa example}
ext0 <- c(0, 15, 0, 10)
set.seed(4236)
# sample 20 random points to subdivide
pts <- vect(spatSample(ext(ext0), 30, lonlat = TRUE))
subexts <- lp_subdiv_pts(pts, 5, ext0)
# bias by resampling a bit in one sub extent
for (i in 1:2) {
    pts <- rbind(pts[-5, ], vect(spatSample(ext(subexts[nrow(subexts) - 1, ]), 5, lonlat = TRUE)))
    subexts <- lp_subdiv_pts(pts, 10, ext0) # subdivide with 30% threshold
}

# generate 3 absences per presence
n <- 3
# one third without subdivision
pts_gen1 <- vect(spatSample(ext(ext0), 1 * nrow(pts), lonlat = TRUE))
# two thirds with subdivision for bias correction
pts_gen2 <- vect(spatSample(ext(ext0), 1, lonlat = TRUE)) # initialize
for (e in seq_len(nrow(subexts))) {
    pts_in <- crop(pts, ext(subexts[e, ]))
    pts_g <- vect(spatSample(ext(subexts[e, ]), 2 * nrow(pts_in), lonlat = TRUE))
    pts_gen2 <- rbind(pts_gen2, pts_g)
}
pts_gen2 <- pts_gen2[-1, ]

# plot
p1 <- ggplot() +
    geom_spatvector(data = pts_gen1, size = 2, shape = 1) +
    geom_spatvector(data = pts_gen2, size = 2, shape = 8) +
    geom_spatvector(data = pts, size = 2, colour = "#00ba38") +
    theme_pubr() +
    theme(axis.text.x = element_blank(), axis.ticks.x = element_blank(), axis.line.x.bottom = element_blank(), axis.line.y.left = element_blank(), axis.text.y = element_blank(), axis.ticks.y = element_blank())
for (i in seq_len(nrow(subexts))) {
    p1 <- p1 + geom_spatvector(data = vect(ext(subexts[i, ])), colour = "#3f80f1", fill = NA, linetype = "solid", linewidth = 0.5)
}

# pa subdiv example with 2008
pa <- readRDS("R/data/occurrence_data/axyridis_pa.rds")
lc_ref <- rast("R/data/cropped_rasters/Cop_LC_2002_eu.tif")

pa <- subset(pa, Area == "eu" & Year == 2008)
pa_v <- vect(pa, geom = c("Lon", "Lat"), crs = crs(lc_ref))

# countries map
countries <- ne_countries(scale = "medium", returnclass = "sf")
countries <- st_transform(countries, crs = crs(lc_ref))
countries <- vect(countries)

p2 <- ggplot(crop(countries, ext(lc_ref))) +
    geom_spatvector(colour = "black", fill = "white") +
    geom_spatvector(data = subset(pa_v, pa_v$Presence == "absent"), colour = "grey", size = 1) +
    geom_spatvector(data = subset(pa_v, pa_v$Presence == "present"), colour = "#00ba38", size = 1) +
    theme_pubr()
subexts <- readRDS("R/data/plotting/axyridis_abs_gen_subexts.rds")
for (i in seq_len(nrow(subexts))) {
    p2 <- p2 + geom_spatvector(data = vect(ext(subexts[i, ]), crs = crs(lc_ref)), colour = "#3f80f1", fill = NA, linetype = "solid", linewidth = 0.5)
}
p <- ggarrange(p1, p2, ncol = 1, labels = c("A", "B"), font.label = list(size = 20))
ggsave(p, width = 6, height = 7, bg = "white", filename = "R/figures/ext-subdiv.png")
```

```{r, combine native niche plot and pca axis}
filepath = "R/plots/niche_comp/as_eu_tot_niche.png"
p1 <- image_read(filepath)
p1 <- image_annotate(p1, "A", color = "black", size = 40, , font = "Helvetica-Bold")
filepath = "R/plots/niche_comp/as_eu_pca.png"
p2 <- image_read(filepath)
p2 <- image_annotate(p2, "B", color = "black", size = 40, , font = "Helvetica-Bold")

img <- image_append(c(p1, p2))
image_write(img, path = "R/figures/as-eu-tot-niche-w-pca.png", format = "png")
```

```{r, combine some niche year plots into one figure}
years <- c("20022003", "20122013", "20212022")
p <- list()
for (i in seq_len(length(years))) {
    p[[i]] <- image_read(paste0("R/plots/niche_comp/single_ys/eu_", years[i], "_niche.png"))
}

img <- image_append(c(p[[1]], p[[2]]))
img <- image_append(c(img, p[[3]]))
image_write(img, path = "R/figures/eu-niche-ys.png", format = "png")
```

```{r, plot of eu niche dynamic indices over time}
# get niche dynamic results + overlap per year
dyn <- as.data.frame(readRDS("R/data/modelling/niche_y_dynamic.rds"))
ol <- readRDS("R/data/modelling/niche_y_overlap.rds")
colnames(ol) <- "overlap"
years <- 2002:2020
dyn <- cbind(years, head(dyn, length(years)), head(ol, length(years)))
# convert to long format
long_dyn <- data.frame()
for (i in 1:4) {
    di <- dyn[, c(1, i + 1)]
    di$index <- colnames(dyn)[i + 1]
    colnames(di)[2] <- "value"
    long_dyn <- rbind(long_dyn, di)
}
long_dyn$index <- factor(long_dyn$index, levels = colnames(dyn)[2:5]) # fix order

p <- ggplot(data = long_dyn, aes(x = years, y = value, color = index, shape = index)) +
    geom_point(size = 2.5) +
    geom_line() +
    scale_color_manual(values = c("#f8766d", "#619cff", "#00ba38", "black")) +
    coord_cartesian(xlim = c(2002, max(years)), ylim = c(0, 1)) +
    theme_pubr() +
    ftheme

ggsave(p, width = 8, height = 5, filename = "R/figures/eu-niche-dyn.png")
```

```{r, create variable table with VIFs}
vifs <- readRDS("R/data/modelling/var_select_vifs.rds")

vars <- data.frame(matrix(ncol = nrow(vifs), nrow = 0))
vars <- rbind(vars, round(vifs[, 1], digits = 2))
colnames(vars) <- as.list(gsub(".*s\\((.+)\\)*.", "\\1", rownames(vifs))) # no s()
name <- "VIF"
vars <- cbind(name, vars)

# write to .csv for table
write.csv(vars, file = "Latex/mainthesis/tab-var-vifs.csv", row.names = FALSE, quote = FALSE)
```

```{r, landcover PCA contribution table}
lc_pca <- readRDS("R/data/modelling/var_select_lc_pca_res.rds")
lc_contrib <- lc_pca$var$contrib
lc_contrib <- format(signif(lc_contrib, digits = 3), scientific = TRUE)
lc_contrib <- cbind(lcc = gsub("lc_", "", rownames(lc_contrib)), e = "", lc_contrib)
lc_var_percent <- c("\\% var", "", round(lc_pca$eig[1:(ncol(lc_contrib)-2), 2], digits = 2))
lc_contrib <- rbind(lc_contrib, lc_var_percent)
colnames(lc_contrib) <- c("Class", "", "{lc}{1}", "{lc}{2}", "{lc}{3}", "{lc}{4}", "{lc}{5}", "{lc}{6}")
# curly braces for center alignment in table (workaround)

write.csv(lc_contrib, file = "Latex/mainthesis/tab-lc-contrib.csv", row.names = FALSE, quote = FALSE)
```

```{r, Cop lcc explanation table}
lccs <- read.csv(file = "R/plots/Cop_LCCS_legend.csv", header = TRUE)

lccs$name <- gsub("_", " ", lccs$name)
lccs$colour <- NULL
colnames(lccs) <- c("\\textbf{Class}", "\\textbf{Explanation}") # error workaround

write.csv(lccs, file = "Latex/mainthesis/tab-Cop-lcc.csv", row.names = FALSE, quote = FALSE)
```

```{r, plot sdm performance over time}
years <- 2002:2020

# get tpr for each year
tpr_fy <- data.frame(year = years) # initialize tpr dataframe
for (y in years) {
    eval <- readRDS(paste0("R/data/modelling/eval_mods/eval_mod_", y, ".rds")) # read eval
    for (m in 1:5) { # get tpr for each model
        # fy tpr df
        ev <- eval[1, ]
        tpr_fy[y - 2001, m + 1] <- ev[[m]]$sensitivity
        colnames(tpr_fy)[m + 1] <- ev[[m]]$model
    }
}
# save tpr results
mod_tpr_res <- cbind(tpr_fy)
saveRDS(mod_tpr_res, file = "R/data/modelling/mod_tpr_res.rds")

# convert to long format
long_fy <- data.frame()
for (m in 1:5) {
    # fy tpr
    tpr_m <- tpr_fy[, c(1, m + 1)]
    tpr_m$model <- colnames(tpr_m)[2]
    colnames(tpr_m)[2] <- "tpr"
    long_fy <- rbind(long_fy, tpr_m)
}

# get tpr for each year with native model
rnt <- readRDS("R/data/modelling/eval_mods/eval_mod_native.rds")
# get tpr for each year
tpr_nt <- data.frame(year = years) # initialize tpr dataframe
for (y in years) {
    eval <- rnt[paste(y), ] # read eval
    for (m in 1:5) { # get tpr for each model
        tpr_nt[y - 2001, m + 1] <- eval[[m]]$sensitivity
        colnames(tpr_nt)[m + 1] <- eval[[m]]$model
    }
}
# save tpr results for native
saveRDS(tpr_nt, file = "R/data/modelling/mod_tpr_res_nat.rds")

# convert to long format
long_nt <- data.frame()
for (m in 1:5) {
    # fy tpr
    tpr_m <- tpr_nt[, c(1, m + 1)]
    tpr_m$model <- colnames(tpr_m)[2]
    colnames(tpr_m)[2] <- "tpr"
    long_nt <- rbind(long_nt, tpr_m)
}

p1 <- ggplot(data = long_fy, aes(x = year, y = tpr, color = model, shape = model)) +
    geom_point(size = 2.5) +
    geom_line() +
    coord_cartesian(xlim = c(2002, max(years)), ylim = c(0, 1)) +
    theme_pubr() +
    ftheme

p2 <- ggplot(data = long_nt, aes(x = year, y = tpr, color = model, shape = model)) +
    geom_point(size = 2.5) +
    geom_line() +
    coord_cartesian(xlim = c(2002, max(years)), ylim = c(0, 1)) +
    theme_pubr() +
    ftheme

p <- ggarrange(p1, p2, nrow = 2, common.legend = TRUE, labels = "AUTO", font.label = list(size = 20))
ggsave(p, width = 8, height = 8, bg = "white", filename = "R/figures/modelling-res.png")
```

```{r, table with model thresholds used}
ths <- data.frame()
for (y in 2002:2020) {
    th_y <- readRDS(paste0("R/data/modelling/eval_mods/eval_mod_", y, ".rds"))
    for (i in 1:5) {
        ths[y - 2001, i] <- th_y[1, ][[i]]$threshold
    }
}
ths <- rbind("-", ths, "-") # add years 2002 and 2022

th_nat <- readRDS("R/data/modelling/eval_mods/eval_mod_native.rds")
for (y in 2002:2022) {
    for (i in 1:5) {
        ths[y - 2001, i + 5] <- th_nat[y - 2001, ][[i]]$threshold
    }
}

ths <- cbind(2002:2022, ths)
colnames(ths) <- c("Year", "glm inv", "gam inv", "brt inv", "max inv", "ens inv", "glm nat", "gam nat", "brt nat", "max nat", "ens nat")
colnames(ths) <- paste0("\\textbf{", colnames(ths), "}") # workaround for bol first line

write.csv(ths, file = "Latex/mainthesis/tab-mod-ths.csv", row.names = FALSE, quote = FALSE)
```

```{r, correlation model performance with pcount or niche dyn}
tpr_res <- readRDS("R/data/modelling/mod_tpr_res.rds")
dyn <- readRDS("R/data/modelling/niche_y_dynamic.rds")
nvar <- as.data.frame(dyn)$stability
nvar <- nvar[1:19]

# get list of total pa points up to year in EU
pa <- readRDS("R/data/occurrence_data/axyridis_pa.rds")
eu <- subset(pa, Area == "eu")
pcount <- c()
for (i in 2002:2022) {
    pcount[i - 2001] <- nrow(subset(eu, Year <= i))
}
# pcount for years used
pcount <- head(pcount, length(nvar))

# compute pearson correlation for each model
cor_res <- c()
for (i in 2:6) {
    tpr_mod <- tpr_res[, i]
    c_ol <- cor.test(tpr_mod, nvar, method = "pearson")
    c_pc <- cor.test(tpr_mod, pcount, method = "pearson")
    c_res_mod <- c(c_ol$estimate, c_ol$p.value, c_pc$estimate, c_pc$p.value)
    cor_res <- cbind(cor_res, c_res_mod)
}
cor_res <- round(cor_res, digits = 3)
rownames(cor_res) <- NULL
colnames(cor_res) <- colnames(tpr_res)[2:6]
cor_res <- cbind(Model = c("Corr. niche stability", "p-value", "Corr. data count", "p-value"), cor_res)

write.csv(cor_res, file = "Latex/mainthesis/tab-mod-cor-res.csv", row.names = FALSE, quote = FALSE)
```

```{r, suitability prediction of 2022 nat vs 2020 mod}
## predict 2020 model
th_data <- readRDS("R/data/modelling/th_data_mods/th_data_m2020_2022.rds")
pa <- readRDS("R/data/occurrence_data/axyridis_pa.rds")
pa_ext <- readRDS("R/data/occurrence_data/axyridis_pa_vals_extracted.rds")
po <- subset(pa, Area == "eu" & Year == 2022 & Presence == "present")
pa <- subset(pa, Area == "eu" & Year == 2022 & Presence == "absent")
lc_ref <- rast("R/data/cropped_rasters/Cop_LC_2002_eu.tif")
pa_ext <- subset(pa_ext, Area == "eu" & Year == 2022 & Presence == "absent")
th_data <- subset(th_data, pres == 0)
ext_join <- data.frame(Lon = pa_ext$Lon, Lat = pa_ext$Lat, Prob_ens = th_data$ens)

pa_m <- left_join(pa, ext_join)
pa_m <- slice_head(pa_m, n = round(nrow(pa_m) / 3))
pa_m <- pa_m[complete.cases(pa_m), ]
pa_m <- pa_m[order(pa_m$Prob_ens), ]
pa_v <- vect(pa_m, geom = c("Lon", "Lat"), crs = crs(lc_ref))
po_v <- vect(po, geom = c("Lon", "Lat"), crs = crs(lc_ref))

# countries map
countries <- ne_countries(scale = "medium", returnclass = "sf")
countries <- st_transform(countries, crs = crs(lc_ref))
countries <- vect(countries)

# plot with suitability value
p1 <- ggplot(crop(countries, ext(lc_ref))) +
    geom_spatvector(colour = "black", fill = "white") +
    geom_spatvector(data = pa_v, aes(color = Prob_ens)) +
    scale_color_viridis() +
    geom_spatvector(data = po_v, size = 0.1) +
    labs(color = "suitability") +
    theme_pubr()

## predict native model
th_data <- readRDS("R/data/modelling/th_data_mods/th_data_nt_2022.rds")
pa <- readRDS("R/data/occurrence_data/axyridis_pa.rds")
pa_ext <- readRDS("R/data/occurrence_data/axyridis_pa_vals_extracted.rds")
po <- subset(pa, Area == "eu" & Year == 2022 & Presence == "present")
pa <- subset(pa, Area == "eu" & Year == 2022 & Presence == "absent")
lc_ref <- rast("R/data/cropped_rasters/Cop_LC_2002_eu.tif")
pa_ext <- subset(pa_ext, Area == "eu" & Year == 2022 & Presence == "absent")
th_data <- subset(th_data, pres == 0)
ext_join <- data.frame(Lon = pa_ext$Lon, Lat = pa_ext$Lat, Prob_ens = th_data$ens)

pa_m <- left_join(pa, ext_join)
pa_m <- slice_head(pa_m, n = round(nrow(pa_m) / 3))
pa_m <- pa_m[complete.cases(pa_m), ]
pa_m <- pa_m[order(pa_m$Prob_ens), ]
pa_v <- vect(pa_m, geom = c("Lon", "Lat"), crs = crs(lc_ref))
po_v <- vect(po, geom = c("Lon", "Lat"), crs = crs(lc_ref))

p2 <- ggplot(crop(countries, ext(lc_ref))) +
    geom_spatvector(colour = "black", fill = "white") +
    geom_spatvector(data = pa_v, aes(color = Prob_ens)) +
    scale_color_viridis() +
    geom_sf(data = po_v, size = 0.1) +
    labs(color = "suitability") +
    theme_pubr()
p <- ggarrange(p1, p2, nrow = 2, common.legend = TRUE, legend = "top", labels = "AUTO", font.label = list(size = 20))
ggsave(p, width = 6, height = 8, bg = "white", filename = "R/figures/2022-mod-pred.png")
```

```{r, time series of prediction plots}
lp_prediction_plot <- function(year) {
    # get data
    th_data <- readRDS(paste0("R/data/modelling/th_data_mods/th_data_m", year, "_2022.rds"))
    pa <- readRDS("R/data/occurrence_data/axyridis_pa.rds")
    pa_ext <- readRDS("R/data/occurrence_data/axyridis_pa_vals_extracted.rds")
    po <- subset(pa, Area == "eu" & Year == year & Presence == "present")
    pa <- subset(pa, Area == "eu" & Year == 2022 & Presence == "absent")
    lc_ref <- rast("R/data/cropped_rasters/Cop_LC_2002_eu.tif")
    pa_ext <- subset(pa_ext, Area == "eu" & Year == 2022 & Presence == "absent")
    th_data <- subset(th_data, pres == 0)
    ext_join <- data.frame(Lon = pa_ext$Lon, Lat = pa_ext$Lat, Prob_ens = th_data$ens)

    pa_m <- left_join(pa, ext_join)
    pa_m <- slice_head(pa_m, n = round(nrow(pa_m) / 3))
    pa_m <- pa_m[complete.cases(pa_m), ]
    pa_m <- pa_m[order(pa_m$Prob_ens), ]
    pa_v <- vect(pa_m, geom = c("Lon", "Lat"), crs = crs(lc_ref))
    po_v <- vect(po, geom = c("Lon", "Lat"), crs = crs(lc_ref))

    # countries map
    countries <- ne_countries(scale = "medium", returnclass = "sf")
    countries <- st_transform(countries, crs = crs(lc_ref))
    countries <- vect(countries)

    # plot with suitability value
    p <- ggplot(crop(countries, ext(lc_ref))) +
        geom_spatvector(colour = "black", fill = "white") +
        geom_spatvector(data = pa_v, aes(color = Prob_ens)) +
        scale_color_viridis() +
        geom_spatvector(data = po_v, size = 0.1) +
        labs(color = "suitability") +
        ggtitle(paste0(year, " model prediction for 2022")) +
        theme_pubr()
    return(p)
}

plot_list <- list()
for (year in 2002:2020) {
    p <- lp_prediction_plot(year)
    plot_list[[year-2001]] <- p
}

plot <- ggarrange(plotlist = plot_list, nrow = 5, ncol = 4, common.legend = TRUE, legend = "top")
ggsave(plot, width = 16, height = 20, bg = "white", filename = "R/figures/2003to2021-mod-pred.png")
```

```{r}
td <- difftime(Sys.time(), tot_time, units = "secs")[[1]]
cat("building and model evaluation finished:", td, "secs", "\n")
```
